+++ lib/screens/register_tab_screen.dart
@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -12,6 +10,7 @@ import 'package:romrom_fe/screens/item_detail_description_screen.dart';
 import 'package:romrom_fe/widgets/common/romrom_context_menu.dart';
 import 'package:romrom_fe/widgets/common/error_image_placeholder.dart';
 import 'package:romrom_fe/widgets/skeletons/register_tab_skeleton.dart';
+import 'package:romrom_fe/widgets/common/glass_header_delegate.dart';
 import 'dart:async';
 import 'package:romrom_fe/utils/common_utils.dart';
 import 'package:romrom_fe/services/apis/item_api.dart';
@@ -238,7 +237,15 @@ class _RegisterTabScreenState extends State<RegisterTabScreen>
                     SliverPersistentHeader(
                       pinned: true,
                       delegate: GlassHeaderDelegate(
-                        toggle: _buildToggleWidget(),
+                        headerTitle: '나의 등록된 물건',
+                        toggle: GlassHeaderToggleBuilder.buildDefaultToggle(
+                          animation: _toggleAnimation,
+                          isRightSelected: _isCompletedSelected,
+                          onLeftTap: () => _onToggleChanged(false),
+                          onRightTap: () => _onToggleChanged(true),
+                          leftText: '판매 중',
+                          rightText: '거래 완료',
+                        ),
                         statusBarHeight:
                             MediaQuery.of(context).padding.top, // ★ 꼭 전달
                         toolbarHeight: 58.h,
@@ -524,89 +531,6 @@ class _RegisterTabScreenState extends State<RegisterTabScreen>
     );
   }
 
-  /// 토글 위젯 (판매 중 / 거래 완료)
-  Widget _buildToggleWidget() {
-    return Padding(
-      padding: EdgeInsets.fromLTRB(24.w, 0, 24.w, 24.h), // 상단 패딩 제거, 하단만 24px
-      child: Container(
-        width: 345.w,
-        height: 46.h,
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(10.r),
-          color: AppColors.secondaryBlack, // #2C2D36
-        ),
-        child: AnimatedBuilder(
-          animation: _toggleAnimation,
-          builder: (context, child) {
-            return Stack(
-              children: [
-                // 애니메이션 선택된 배경
-                Positioned(
-                  left: 2.w +
-                      (_toggleAnimation.value * 171.w), // 2px + 170px + 1px gap
-                  top: 2.h,
-                  child: Container(
-                    width: 170.w,
-                    height: 42.h,
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(8.r),
-                      color: AppColors.primaryBlack, // #1D1E27
-                    ),
-                  ),
-                ),
-                // 텍스트 버튼들
-                Row(
-                  children: [
-                    Expanded(
-                      child: GestureDetector(
-                        onTap: () => _onToggleChanged(false),
-                        child: Container(
-                          height: 46.h,
-                          color: Colors.transparent,
-                          alignment: Alignment.center,
-                          child: AnimatedDefaultTextStyle(
-                            duration: const Duration(milliseconds: 300),
-                            style: CustomTextStyles.p1.copyWith(
-                              color: !_isCompletedSelected
-                                  ? AppColors.textColorWhite
-                                  : AppColors.opacity60White,
-                              fontWeight: FontWeight.w500,
-                            ),
-                            child: const Text('판매 중'),
-                          ),
-                        ),
-                      ),
-                    ),
-                    Expanded(
-                      child: GestureDetector(
-                        onTap: () => _onToggleChanged(true),
-                        child: Container(
-                          height: 46.h,
-                          color: Colors.transparent,
-                          alignment: Alignment.center,
-                          child: AnimatedDefaultTextStyle(
-                            duration: const Duration(milliseconds: 300),
-                            style: CustomTextStyles.p1.copyWith(
-                              color: _isCompletedSelected
-                                  ? AppColors.textColorWhite
-                                  : AppColors.opacity60White,
-                              fontWeight: FontWeight.w500,
-                            ),
-                            child: const Text('거래 완료'),
-                          ),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ],
-            );
-          },
-        ),
-      ),
-    );
-  }
-
   /// 토글 상태 변경
   void _onToggleChanged(bool isCompleted) {
     if (_isCompletedSelected != isCompleted) {
@@ -770,136 +694,5 @@ class _RegisterTabScreenState extends State<RegisterTabScreen>
         );
       }
     }
-  }
-}
-
-class GlassHeaderDelegate extends SliverPersistentHeaderDelegate {
-  final Widget toggle;
-  final double statusBarHeight;
-  final double toolbarHeight; // 58.h
-  final double toggleHeight; // 70.h
-  final double expandedExtra; // 큰 제목/여백 등 “펼침 전용” 추가 높이
-  final Color tintBase;
-  final bool enableBlur;
-
-  GlassHeaderDelegate({
-    required this.toggle,
-    required this.statusBarHeight,
-    required this.toolbarHeight,
-    required this.toggleHeight,
-    this.expandedExtra = 32.0, // 큰 제목 여백 등 (원하는 만큼)
-    this.tintBase = Colors.black,
-    this.enableBlur = true,
-  }) : assert(statusBarHeight >= 0 && toolbarHeight >= 0 && toggleHeight >= 0);
-
-  // ⬇️ 토글을 포함해서 최소 높이를 정의 → 토글이 항상 보임
-  @override
-  double get minExtent => statusBarHeight + toolbarHeight + toggleHeight;
-
-  // ⬇️ 펼쳐질 때만 추가로 커지는 영역(큰 제목 등)
-  @override
-  double get maxExtent => minExtent + expandedExtra;
-
-  @override
-  Widget build(
-      BuildContext context, double shrinkOffset, bool overlapsContent) {
-    final extraRange = (maxExtent - minExtent).clamp(0.0, double.infinity);
-    final t =
-        extraRange == 0 ? 1.0 : (shrinkOffset / extraRange).clamp(0.0, 1.0);
-
-    final sigma = enableBlur ? lerpDouble(0, 30, t)! : 0.0;
-
-    return ClipRect(
-      child: Stack(
-        fit: StackFit.expand,
-        children: [
-          // 1) 블러
-          BackdropFilter(
-            filter: ImageFilter.blur(sigmaX: sigma, sigmaY: sigma),
-            child: const SizedBox.expand(),
-          ),
-
-          // 2) 틴트(사파리 감성은 그라데이션 추천)
-          Container(
-            decoration: const BoxDecoration(
-              color: AppColors.opacity90PrimaryBlack,
-            ),
-          ),
-
-          // 3) 큰 제목(펼침에서만 보이고 스크롤되면 사라짐)
-          Positioned(
-            left: 24,
-            right: 24,
-            top: statusBarHeight + 32,
-            child: Opacity(
-              opacity: 1.0 - t,
-              child: Text('나의 등록된 물건', style: CustomTextStyles.h1),
-            ),
-          ),
-
-          // 4) 작은 제목(툴바 타이틀 역할) — 스크롤될수록 나타남
-          Positioned(
-            left: 0,
-            right: 0,
-            top: statusBarHeight,
-            height: toolbarHeight,
-            child: IgnorePointer(
-              ignoring: true,
-              child: Center(
-                child: Opacity(
-                  opacity: t,
-                  child: Padding(
-                    padding: const EdgeInsets.only(top: 16, bottom: 24),
-                    child: Text(
-                      '나의 등록된 물건',
-                      style: CustomTextStyles.h3
-                          .copyWith(fontWeight: FontWeight.w600),
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ),
-
-          // 5) 🔒 토글: 항상 보이는 영역(최소 높이에 포함시켰기 때문에 사라지지 않음)
-          Positioned(
-            left: 0,
-            right: 0,
-            top: statusBarHeight + toolbarHeight + lerpDouble(24, 0, t)!,
-            height: toggleHeight,
-            child: Material(
-              color: Colors.transparent,
-              child: toggle,
-            ),
-          ),
-
-          // 6) 하단 라인(살짝)
-          Positioned(
-            left: 0,
-            right: 0,
-            bottom: 0,
-            child: Opacity(
-              opacity: 0.15 * t,
-              child: const Divider(
-                height: 1,
-                thickness: 1,
-                color: AppColors.opacity20Black,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  @override
-  bool shouldRebuild(covariant GlassHeaderDelegate old) {
-    return toggle != old.toggle ||
-        statusBarHeight != old.statusBarHeight ||
-        toolbarHeight != old.toolbarHeight ||
-        toggleHeight != old.toggleHeight ||
-        expandedExtra != old.expandedExtra ||
-        enableBlur != old.enableBlur ||
-        tintBase != old.tintBase;
   }
 }
